Edited

+++ MCAL/std_types.h
@@ -36,10 +36,10 @@
 typedef unsigned char boolen;
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned int uint32;
+typedef unsigned long uint32;
 typedef signed char sint8;
 typedef signed short sint16;
-typedef signed int sint32;
+typedef signed long sint32;
 typedef float float32;
 
 typedef uint8 Std_ReturnType;
